+++ src/lschfb.f90
@@ -186,7 +186,7 @@ subroutine lschfb(nn,ll,ierr,ee,rr,vv,uu,up,zz,mmax,mch,srel)
 ! start inward integration at 10*classical turning
 ! point with simple exponential
 
-         nin=mch+2.3d0/al
+         nin=int(mch+2.3d0/al)
          if(nin+4>mmax) nin=mmax-4
          xkap=dsqrt(sls/rr(nin)**2 + 2.0d0*(vv(nin)-ee))
