+++ src/lschpsbar.f90
@@ -162,7 +162,7 @@ subroutine lschpsbar(nn,ll,ierr,ee,emin,emax,rr,vv,uu,up,mmax,mbar,tht)
 ! start inward integration at 10*classical turning
 ! point with simple exponential
 
-         nin=mch+2.3d0/al
+         nin=int(mch+2.3d0/al)
          if(nin<mbar-4) then
             if(nin+4>mmax) nin=mmax-4
             xkap=dsqrt(sls/rr(nin)**2 + 2.0d0*(vv(nin)-ee))
